+++ flutter/example/lib/main.dart
@@ -23,9 +23,11 @@ final _channel = const MethodChannel('example.flutter.sentry.io');
 Future<void> main() async {
   await setupSentry(() => runApp(
         SentryScreenshotWidget(
-          child: DefaultAssetBundle(
-            bundle: SentryAssetBundle(enableStructuredDataTracing: true),
-            child: MyApp(),
+          child: SentryUserInteractionWidget(
+            child: DefaultAssetBundle(
+              bundle: SentryAssetBundle(enableStructuredDataTracing: true),
+              child: MyApp(),
+            ),
           ),
         ),
       ));
@@ -41,6 +43,10 @@ Future<void> setupSentry(AppRunner appRunner) async {
     options.attachThreads = true;
     options.enableWindowMetricBreadcrumbs = true;
     options.addIntegration(LoggingIntegration());
+    options.sendDefaultPii = true;
+    options.reportSilentFlutterErrors = true;
+    options.enableNdkScopeSync = true;
+    options.enableUserInteractionTracing = true;
     options.attachScreenshot = true;
     // We can enable Sentry debug logging during development. This is likely
     // going to log too much for your app, but can be useful when figuring out
@@ -124,6 +130,7 @@ class MainScaffold extends StatelessWidget {
             ),
             ElevatedButton(
               onPressed: () => tryCatch(),
+              key: Key('dart_try_catch'),
               child: const Text('Dart: try catch'),
             ),
             ElevatedButton(
@@ -201,7 +208,8 @@ class MainScaffold extends StatelessWidget {
               child: const Text('Flutter: Load assets'),
             ),
             ElevatedButton(
-              onPressed: () => makeWebRequestWithDio(context),
+              key: Key('dio_web_request'),
+              onPressed: () async => await makeWebRequestWithDio(context),
               child: const Text('Dio: Web request'),
             ),
             ElevatedButton(
@@ -560,6 +568,7 @@ Future<void> makeWebRequestWithDio(BuildContext context) async {
   dio.addSentry(
     captureFailedRequests: true,
     maxRequestBodySize: MaxRequestBodySize.always,
+    maxResponseBodySize: MaxResponseBodySize.always,
   );
 
   final transaction = Sentry.getSpan() ??
@@ -568,16 +577,20 @@ Future<void> makeWebRequestWithDio(BuildContext context) async {
         'request',
         bindToScope: true,
       );
+  final span = transaction.startChild(
+    'dio',
+    description: 'desc',
+  );
   Response<String>? response;
   try {
     response = await dio.get<String>('https://flutter.dev/');
-    transaction.status = SpanStatus.ok();
+    span.status = SpanStatus.ok();
   } catch (exception, stackTrace) {
-    transaction.throwable = exception;
-    transaction.status = SpanStatus.internalError();
+    span.throwable = exception;
+    span.status = SpanStatus.internalError();
     await Sentry.captureException(exception, stackTrace: stackTrace);
   } finally {
-    await transaction.finish();
+    await span.finish();
   }
 
   await showDialog<void>(