+++ flutter/lib/src/sentry_flutter_options.dart
@@ -188,6 +188,20 @@ class SentryFlutterOptions extends SentryOptions {
   /// The [SentryScreenshotWidget] has to be the root widget of the app.
   bool attachScreenshot = false;
 
+  /// Enable or disable automatic breadcrumbs for User interactions Using [Listener]
+  ///
+  /// Requires adding the [SentryUserInteractionWidget] to the widget tree.
+  /// Example:
+  /// runApp(SentryUserInteractionWidget(child: App()));
+  bool enableUserInteractionBreadcrumbs = true;
+
+  /// Enables the Auto instrumentation for user interaction tracing.
+  ///
+  /// Requires adding the [SentryUserInteractionWidget] to the widget tree.
+  /// Example:
+  /// runApp(SentryUserInteractionWidget(child: App()));
+  bool enableUserInteractionTracing = false;
+
   @internal
   late RendererWrapper rendererWrapper = RendererWrapper();